Use urandom only on Linux. WIP more tests covering pipes are needed. We need to test epd files with keys and without passwords

+++ func_tests/run_all_tests.sh
@@ -10,7 +10,14 @@ fi
 
 BIN="$1"
 TMP_DIR="./tmp"
-PLAIN_TEXT_FILE=$TMP_DIR/random.dat
+OS=$(uname)
+if [ "$OS" = "Linux" ]
+then
+    PLAIN_TEXT_FILE=$TMP_DIR/random.dat
+else
+    PLAIN_TEXT_FILE=plain_text.txt
+fi
+
 export PLAIN_TEXT_FILE
 
 rm -Rf $TMP_DIR
@@ -18,7 +25,10 @@ rm -Rf ./epd_encrypted_last
 rm -Rf ./gpg_encrypted_last
 
 mkdir -p $TMP_DIR
-dd if=/dev/urandom of=$PLAIN_TEXT_FILE bs=1M count=15
+if [ "$OS" = "Linux" ]
+then
+    dd if=/dev/urandom of=$PLAIN_TEXT_FILE bs=1M count=15
+fi
 
 # diffrent key file and passphrase combinations
 ./epd_encryption_test.sh $BIN "$PLAIN_TEXT_FILE"
